+++ build-support/githooks/pre-commit
@@ -19,12 +19,10 @@ DIRS_TO_CHECK=(
   pants-plugins
   examples
   contrib
-  # TODO(6071): Add `build-support/bin` once we update the check_header_helper.py script to
-  # allow Python 3 headers (i.e. without __future__ and coding=utf-8).
+  build-support/bin
 )
 
-# TODO(#7068): Fix all these checks to only act on staged files with
-# `git diff --cached --name-only`!
+# TODO(#7068): Fix all these checks to only act on staged files using $ADDED_FILES!!
 
 # TODO: test all the scripts in this file in test_git_hooks.py, remove uses of `|| exit 1`, and add an
 # integration test!
